// data/test-data/credentials-values.enum.ts
export enum Credentials {
    USER_EMAIL = "test123@test.com",
    PASSWORD = "12345678Aa",
    USER_ID = "aa 22-33-12 a"
}

export enum BlockedUser {
    EMAIL = "ndn@gmail.com",
    PASSWORD = "FgMC9fUNTH2Jh55",
    ID = "555555",
}

export enum RecoveryPasswordUser{
    USER_EMAIL = "test2a@gmail.com",
    USER_ID = "bb-22-33-13-a",
}

export enum UnregisteredUser {
    EMAIL = "aa@trt.com",
    PASSWORD = "asdPS112Vs",
    ID = "XF-12-99-43-A",
}