// data/test-data/css-values.enum.ts
export enum CssValues {
    YELLOW_DASHED_HEADER = "rgb(255, 214, 0)",
    YELLOW_DASHED_FOOTER = "rgb(237, 199, 0)",
    BLUE_UNDERLINE = "underline solid rgb(64, 84, 178)",
    BLUE_DEFAULT = "none rgb(64, 84, 178)",
    TRANSPARENT_BORDER = "rgba(0, 0, 0, 0)",
    YELLOW_COLLOR = "rgb(255, 214, 0)",
    GRAY_BACKGROUND = "rgb(238, 238, 238)",
    GRAY_COLLOR = "rgb(238, 238, 238)",
    YELLOW_BORDER_BOTTOM = "2px solid rgb(255, 214, 0)",
    RED_COLOR = "rgb(242, 69, 61)",
    GREY_BORDER = "1px solid rgb(158, 158, 158)",
    LIGHT_GREY_BORDER = "1px solid rgb(238, 238, 238)",
}